requirements.txt
openai
openai-whisper
pythonaudio